Add LaneD thread. COmpute D3_HiProt binary map (h>hcrit). Fix missing user defined heightThreshold param parsing section

+++ include/lad_processing.hpp
@@ -27,15 +27,14 @@ using namespace cv;  // OpenCV
 namespace lad
 { // landing area detection algorithm namespace
     typedef struct parameterStruct_{
-        float robotHeight;
-        float robotWidth;
-        float robotLength;
-        float rotation;
-        
-        float heightThreshold;
-        float slopeThreshold;
-        float groundThreshold;
-        float protrusionSize;
+        double robotHeight;
+        double robotWidth;
+        double robotLength;
+        double rotation;
+        double heightThreshold;
+        double slopeThreshold;
+        double groundThreshold;
+        double protrusionSize;
 
         float alphaShapeRadius;
     }parameterStruct;
